Fixed an DeprecationWarning:
.../erl_eventhandler.py:332: DeprecationWarning: integer argument expected, got float
  return _GetTk().createtimerhandler(round(timeLeft * 1000), cb)

It turns out round(1.2 * 1000)  --> 1200.0 (a float!)
while    int(round(1.2 * 1000)) --> 1200   (an integer)

This was not what I expected (though perhaps a bit more consequent?)
(Tested with python 2.4.2 (on gentoo) and python 2.3.4 (on Debian))

+++ erl_eventhandler.py
@@ -329,7 +329,7 @@ Note: When using the Tkinter eventhandler, you cannot delete timer-events."""
             self.timerEvents.sort()
             return newTimerEvent.id
         elif self.state == self.STATE_TK:
-            return _GetTk().createtimerhandler(round(timeLeft * 1000), cb)
+            return _GetTk().createtimerhandler(int(round(timeLeft * 1000)), cb)
 
     def DelTimerEvent(self, id):
         """Unregister a timer callback.
